+++ frontend/src/components/Cart.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { useMutation } from '@apollo/client';
 import { useMenu } from '@/utils/useMenu';
 import { rubikDoodleShadow } from '@/styles/fonts';
 import formatMoney from '@/utils/formatMoney';
@@ -8,11 +9,18 @@ import { useUser } from './User';
 
 import styles from './cart.module.scss';
 
+import GET_CURRENT_USER from '../gql/getCurrentUser.gql';
+import REMOVE_PRODUCT_FROM_CART from '../gql/removeProductFromCart.gql';
+
 export default function Cart() {
   const {
     isCartOpen,
     closeCart,
   }: any = useMenu();
+
+  const [removeProduct, { data, loading, error }] = useMutation(REMOVE_PRODUCT_FROM_CART, {
+    refetchQueries: [{ query: GET_CURRENT_USER }],
+  });
 
   function getCartTotal(cart: any) {
     return cart.reduce((total:any, cartItem:any) => {
@@ -26,6 +34,14 @@ export default function Cart() {
     }, 0);
   }
 
+  function removeItemFromCart(e, cartItem: any) {
+    removeProduct({
+      variables: {
+        where: { id: cartItem.id }
+      }
+    });
+  }
+  
   const user = useUser();
 
   const { cart = [] } = user || {};
@@ -45,6 +61,14 @@ export default function Cart() {
       <ul className={styles.cartItems}>
         {cart.length >= 1 && cart.map((cartItem:any) => (
           <li className={styles.cartItem} key={cartItem.id}>
+            <button
+              type="button"
+              onClick={(e) => removeItemFromCart(e, cartItem)}
+              className={styles.removeButton}
+              disabled={loading}
+            >
+              <img src="/trash-can.png" alt="Remove" />
+            </button>
             <span className={styles.itemTitle}>{cartItem.product.title} x {cartItem.quantity}</span>
             <span className={styles.itemPrice}>{formatMoney(cartItem.product.price)}</span>
           </li>