+++ frontend/src/components/ProductsDisplay.tsx
@@ -1,3 +1,7 @@
+'use client'
+
+import Link from 'next/link';
+
 import formatMoney from '@/utils/formatMoney';
 import { useMutation } from '@apollo/client';
 import { useUser } from './User';
@@ -13,20 +17,15 @@ export default function ProductsDisplay({
   data: Array<any>,
 }) {
   const [addItem, { data: proudctData, loading, error }] = useMutation(ADD_PRODUCT_TO_CART);
-
   const user = useUser();
 
-  async function addItemToCart(e: any, product: any) {
-    e.preventDefault();
-
+  async function addItemToCart(product: any) {
     const res = await addItem({
       variables: {
         productId: product.id,
       },
       refetchQueries: [{ query: GET_CURRENT_USER }]
     });
-
-    // TODO this works, should open the cart or something to show the user it worked
   }
 
   return (
@@ -60,16 +59,28 @@ export default function ProductsDisplay({
                 <span key={tag.name}>{tag.name}</span>
               ))}
             </p>
-            <div className={styles.buttonWrapper}>
-              <fieldset disabled={loading}>
-                <button type="button">Edit</button>
-                <button type="button" onClick={(e) => addItemToCart(e, product)}>Add</button>
-                <button type="button">Delete</button>
-              </fieldset>
-            </div>
+              <div className={`${styles.buttonWrapper} ${!user ? styles.loggedOut : ''}`}>
+                  <fieldset disabled={loading}>
+                    {user && (
+                      <>
+                        <button type="button">Edit</button>
+                        <button type="button" onClick={() => addItemToCart(product)}>Add</button>
+                        <button type="button">Delete</button>
+                      </>
+                    )}
+                    {!user && (
+                      <Link
+                        href="/signin"
+                        className="button"
+                      >
+                        Sign in to purchase
+                      </Link>
+                    )}
+                  </fieldset>
+              </div>
           </div>
-        )
-      })}
+        )})
+      }
     </div>
   );
 }