+++ frontend/src/utils/useMenu.tsx
@@ -37,6 +37,7 @@ function MenuStateProvider({ children }: any) {
   }
 
   function openCart() {
+    console.log('this ran');
     setIsCartOpen(true);
   }