+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.6] - 2024-02-20
+### Added
+- Added cart count to navigation menu
+- Added ability to remove cart items from frontend
+
+### Updated
+- Updated styling of a few navigation related items
+- Cleaned up mobile view of css for cart
+- Update styling of signin form
+
 ## [0.2.5] - 2024-02-20
 ### Added
 - Added ability to add cart item from frontend